Selfplay/change delta threshold ranking 1.0

+++ source/give_clue.py
@@ -1,6 +1,7 @@
 import gensim
 import itertools
 import numpy as np
+from sklearn.metrics.pairwise import cosine_similarity
 import pickle
 import os
 import csv
@@ -13,10 +14,16 @@ import sys
 class Clue(object):
     """ Class for a clue with its score and ranked (card, score) answer pairs."""
 
-    def __init__(self, clue, sorted_card_score_pairs, team):
+    def __init__(self, clue, sorted_card_score_pairs, delta, penalize_negative, normalize_negative, alpha, team, logger):
         self.clue = clue
         self.sorted_card_score_pairs = sorted_card_score_pairs    
         self.team = team
+        self.logger = logger
+        self.delta = delta
+        self.penalize_negative = penalize_negative
+        self.normalize_negative = normalize_negative
+        self.alpha = alpha
+        self.cropped_threshold = None
         self.total_score, self.clue_number = self._calculate_score_with_threshold()
 
     def _calculate_score_with_threshold(self):
@@ -24,24 +31,60 @@ class Clue(object):
         Helper function.
 
         total_score = sum_{card in clue_answers} similarity(clue, card).
-        clue_answers set is defined as all the cards with similarity greater than 
+        clue_answers set is defined as all the cards with similarity greater than
         for any negative card (wrong team, assasin, normal).
       
         :return: total_score, clue_number
         """
-        # TODO: compute normalized negative score and add it to the total score.
-        
+
         clue_number = 0
+        positive_score, negative_score = 0, 0
+        negative_number = 0
         total_score = 0
-    
-        for card, score in self.sorted_card_score_pairs:
-            # Collect positive set until the first negative word occurrence.
-            if card.color in [self.team, "DOUBLE"]:
-                clue_number += 1
-                total_score += score
-            else:
+
+        # find largest negative score
+        largest_negative_score = -1.
+        for ix, (card, score) in enumerate(self.sorted_card_score_pairs):
+            # find maximum score of negative word
+            if card.color not in [self.team, "DOUBLE"]:
+                largest_negative_score = score
                 break
-                
+
+        # add scores higher than threshold + largest negative score to positive_score
+        for card, score in self.sorted_card_score_pairs:
+           if (score > (self.delta+largest_negative_score)
+               and card.color in [self.team, "DOUBLE"]):
+               clue_number += 1
+               positive_score += score
+           elif card.color not in [self.team, "DOUBLE"]:
+               negative_score += score
+               negative_number += 1
+           else:
+               continue
+
+        if not self.penalize_negative:
+            self.logger.info("negative score set to 0.")
+            negative_score = 0
+
+        # if threshold(delta) is large, there will be no clues.
+        # try to give at least one clue
+        # select the positive card with score larger than largest_negative_score.
+        if clue_number == 0:
+            self.logger.debug("clue number: 0.")
+            for card, score in self.sorted_card_score_pairs:
+                if card.color in [self.team, "DOUBLE"]:
+                    positive_score = score
+                    clue_number += 1
+                    self.cropped_threshold = score - largest_negative_score
+                else:
+                    positive_score = 0
+                break
+
+        if self.normalize_negative:
+            total_score = (1-self.alpha) * positive_score - self.alpha * negative_score / negative_number
+        else:
+            total_score = (1-self.alpha) * positive_score - self.alpha * negative_score
+        self.logger.debug("word: {}, positive_score: {}, negative_score: {}, total_score: {}".format(self.clue, positive_score, negative_score, total_score))
         return total_score, clue_number
     
     def get_summary(self):
@@ -76,9 +119,13 @@ class Spymaster(object):
     :param logger: logger for the spymaster.
     """
 
-    def __init__(self, field, embeddings, vocabulary_path, similarities_table_path, logger):
-        
-        self.similarities_table_path = similarities_table_path
+    def __init__(self, field, embeddings, vocabulary_path, keep_similarities_table, similarities_table_path,
+                 logger, game_count):
+
+        self.game_count = game_count
+        self.keep_similarities_table = keep_similarities_table
+        if self.keep_similarities_table:
+            self.similarities_table_path = similarities_table_path
         self.vocabulary_path = vocabulary_path
         self.field = field
         self.logger = logger     
@@ -136,14 +183,14 @@ class Spymaster(object):
                 with open(self.vocabulary_path, 'r') as fin:
                     for line in fin:
                         word = line.strip()
-                        self.update_vocab(word)      
+                        self.update_vocab(word)
             else:
                 raise ValueError("Unknown file format for filter spymaster vocabulary.") 
         else:
             self.logger.info("Load spymaster vocabulary from gensim.models.KeyedVectors.")
             self.vocab = self.model.vocab
             self.vocab_size = len(self.vocab)
-                    
+
         self.logger.info("Spymaster vocabulary size is {}".format(self.vocab_size))
   
 
@@ -155,10 +202,11 @@ class Spymaster(object):
         
         similarities_table_path = self.similarities_table_path
 
-        if similarities_table_path and os.path.exists(similarities_table_path):
+        if self.keep_similarities_table and similarities_table_path and os.path.exists(similarities_table_path):
+            # raise ValueError("self.similarities_table_path is deprecated now. Do not give similarities_table value.")
             self.logger.info(similarities_table_path + " exists. Loading.")
             with open(similarities_table_path, 'rb') as r:
-                self.similarities_table_path = pickle.load(r)
+                self.similarities_table = pickle.load(r)
 
         else:
             self.logger.info("Creating similarities_table from scratch.")
@@ -171,16 +219,23 @@ class Spymaster(object):
                         self.logger.warning("OOV word or card :{}, setting similarity to 0.".format(word))
                         self.similarities_table[w_ix][c_ix] = 0.0
                     else:
-                        self.similarities_table[w_ix][c_ix] = self.model.similarity(word, card.name)
+                        w_vec = self.model[word].reshape(-1, len(self.model[word]))
+                        c_vec = self.model[card.name].reshape(-1, len(self.model[card.name]))
+                        similarity = np.asscalar(cosine_similarity(w_vec, c_vec))
+                        self.similarities_table[w_ix][c_ix] = similarity
+                        # self.similarities_table[w_ix][c_ix] = self.model.similarity(word, card.name)
             
-            if (similarities_table_path):    
+            if similarities_table_path and self.keep_similarities_table:
+                if os.path.exists(similarities_table_path):
+                    raise ValueError("{} already exists, but tries to save it.".format(similarities_table_path))
                 with open(similarities_table_path, 'wb') as w:
-                    pickle.dump(self.similarities_table_path, w)
+                    self.logger.info("saveing table to {}...".format(similarities_table_path))
+                    pickle.dump(self.similarities_table, w)
 
-    def give_clue_with_threshold(self, team, turn_count, top_to_print=5):
+    def give_clue_with_threshold(self, team, turn_count, delta, penalize_negative, normalize_negative, alpha, top_to_print=5):
         """
         Give a clue, maximizig the sum of similarities to the positive words set
-        and minimizing the average similarity to all negative words of the field.
+        and minimizing the average snimilarity to all negative words of the field.
         
         The positive set is determined by top-k words in the predicted ranking
         of answers for the given clue; k is the largest possible value, 
@@ -196,6 +251,7 @@ class Spymaster(object):
             raise ValueError("Team string must be RED or BLUE.")
        
         clue_candidates = []
+        is_hacky_clue = False
         for clue in self.vocab:
             card_score_pairs = []
             clue_ix = self.vocab[clue].index
@@ -204,13 +260,29 @@ class Spymaster(object):
                 card_score_pairs.append((card, score))
 
             sorted_card_score_pairs = sorted(card_score_pairs, key=lambda x: x[1], reverse=True)
-            clue = Clue(clue, sorted_card_score_pairs, team)
+            clue = Clue(clue, sorted_card_score_pairs, delta, penalize_negative, normalize_negative, alpha, team, self.logger)
             clue_candidates.append(clue)
-            
-        clue_candidates = sorted(clue_candidates, key=lambda x: x.total_score, reverse=True)
-        clue = clue_candidates[0]
-        
-        for clue in clue_candidates[:top_to_print]:
-            self.logger.info(clue.get_summary())
 
-        return clue.clue, clue.clue_number, clue.sorted_card_score_pairs
+
+        clue_candidates = sorted(clue_candidates, key=lambda x: x.total_score, reverse=True)
+
+        # exclude the clue whose clue_num is equal to 0 .
+        clue_candidates = list(filter(lambda x: x.clue_number > 0, clue_candidates))
+
+        # find the clue whose threshold is equal to (more than) hparam: alpha.
+        uncropped_clue_candidates = list(filter(lambda x: x.cropped_threshold is None, clue_candidates))
+
+        # if there was no clue (it happens when threshold is too high), pick hacky clues.
+        if len(uncropped_clue_candidates) == 0:
+            clue = list(filter(lambda x: x.cropped_threshold is not None, clue_candidates))[0]
+            is_hacky_clue = True
+            self.logger.info("none of the clue similarity has exceeded threshold. clue: {} has chosen,".format(clue.clue))
+
+        else:
+            clue = uncropped_clue_candidates[0]
+
+        # for clue in clue_candidates[:top_to_print]:
+        #     self.logger.info(clue.get_summary())
+        self.logger.info(clue.get_summary())
+
+        return clue, is_hacky_clue
